% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query - SQL ODBC}
\usage{
query(site, start_date, end_date, variables, password)
}
\arguments{
\item{site}{Site name string like "Altoona". Must have id entry inside function.}

\item{start_date}{start date string like "09-01-2017"}

\item{end_date}{end date string like "09-01-2017"}

\item{variables}{comma separated string like str_c("eq_id", "Timestamp", "UnitKw", "OATemp", sep = ", ")}

\item{password}{the database password}
}
\value{
A dataframe resulting from the SQL query
}
\description{
After setting up an open database connection (ODBC), this function calls the Transformative 
Wave server. Uses the ROBDC package.
}
\examples{
vars <- str_c("eq_id", "Timestamp", "UnitKw", "OATemp", sep = ", ")
data <- query("Altoona", "09-01-2017", "09-02-2017", vars) 

}
\keyword{SQL}
